pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "gunicorn>=25.3.0",
    "python-dotenv>=1.2.2",
    "psycopg2-binary>=2.9.9",
    "validators>=0.22.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]
[tool.uv]
dev-dependencies = [
    "ruff>=0.9.7",
]